+++ CHANGELOG.md
@@ -5,10 +5,12 @@
 - **BUGFIX:** Suchfeld wird nach Enter-Taste zum Start der Suche wieder fokussiert.
 - **BUGFIX:** Programm stürzt nicht mehr ab wenn man versucht während der Indizierung die Blacklist auszuschalten.
 - **BUGFIX(Windows/Linux):** Zeilenhöhe der Tabellen wird nun nach Änderung der Schriftart und -größe korrekt justiert.
+- **BUGFIX:** Link zur Online-Hilfe verweist nurn zur Anleitung, nicht mehr zu den FAQs.
 - **FEATURE:** In der Filmliste kann die Darstellung der Info-Icons nun so konfiguriert werden, dass sie links oder Rechts des Titels dargestellt werden. Dies erfolgt mit der Checkbox unter *Einstellungen/Allgemein->Info-Icons der Listen rechts darstellen*.
 - **FEATURE(Windows/Linux):** Geänderte Programmschriftart und -größe wird nun gespeichert und beim Start wiederhergestellt.
 - **FEATURE:** Die URLs des selektierten Films können nun für HQ und normale Auflösung mittels `Strg+Alt+Shift+H` bzw. `Strg+Alt+Shift+N` oder wie bisher über das Kontextmenü in die Zwischenablage kopiert werden. Unter macOS ersetze `Strg` mit `Command`.
 - **FEATURE:** MV kann mittels `-dq` oder `--download-quit` Parameter zum automatischen Download aller Filme in der Downloadansicht mit anschließender automatischer Beendigung aufgefordert werden.
+- **FEATURE:** Im Hilfe-Menü wurde ein Eintrag zu den Frequently Asked Questions (FAQ) ergänzt.
 
 # **14.0.0**
 - Es wird nun Java 20 verwendet.